// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  USER: 'USER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  BLOCKED: 'BLOCKED',
  DELETED: 'DELETED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const IdeaStatus = {
  DRAFT: 'DRAFT',
  UNDERREVIEW: 'UNDERREVIEW',
  ACCEPTED: 'ACCEPTED',
  REJECTED: 'REJECTED'
} as const

export type IdeaStatus = (typeof IdeaStatus)[keyof typeof IdeaStatus]


export const VoteType = {
  UPVOTE: 'UPVOTE',
  DOWNVOTE: 'DOWNVOTE'
} as const

export type VoteType = (typeof VoteType)[keyof typeof VoteType]


export const PaymentStatus = {
  PAID: 'PAID',
  UNPAID: 'UNPAID',
  PENDING: 'PENDING'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]